requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.10.0
mne>=1.4.0
mne-bids>=0.12.0
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.21.0
Pillow>=9.5.0
tqdm>=4.65.0
h5py>=3.9.0
imageio>=2.31.0
openneuro-py>=2023.1.0